pyproject.toml
[tool.poetry]
name = "dj-git-crawler"
version = "0.1.0"
description = "git indexer"
authors = ["Li Lin <guru.lin@gmail.com>"]
license = "ASL"

[tool.poetry.dependencies]
python = "^3.8"
django = "~3.0.9"
python-dotenv = "^0.14.0"
celery = {version = "~4.4", extras = ["redis"]}
django-celery-results = "^1.2.1"
django-celery-beat = "^2.0.0"
flower = "^0.9.5"
psycopg2 = "^2.8.6"
python-gitlab = "^2.5.0"
PyDriller = "^1.15.2"

[tool.poetry.dev-dependencies]
black = "20.8b1"
flake8 = "^3.8.4"
pytest-django = "^3.10.0"
pytest-dependency = "^0.5.1"
pytest-django-ordering = "^1.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
